New code for register and login forms.
New code in the register and login forms html for the form and inputs. New code in register.mjs and login.mjs for the forms to work. Created a new validation.mjs where the validation code rules for the inputs is done. Some new scss code for the input errors.