New Html, JS and Scss code
New code in index.mjs to get all posts, profile.mjs to get list of profiles and specific.mjs to get the specific post on page. Styling done for them in their own scss files. Added profile link in nav in about page. And added a div in index, profile and specific.html to get the api on the site.