Code for search input/button.
Created a new Api, Handlers and Views folders. And inside these folders I created new files. All these files is made to make the search input and button work to search for lists from the api with a specefic tag(s). Inside the index.html i added a new id for the form, also changed the name to the id of the input.