New code - view bids on page.
In the specific.mjs file I added a new function and also a div to view bids made on a listing. Also added [] to the meida in the create file, because media is an array. Also fixed the media.value. In the list.mjs added item.tags rather than id.